refactor: move observing comms to communication folder
